+++ packages/storybook/package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "lint": "eslint . --ext .js,.ts",
+    "lint": "eslint . --ext .js,.ts 'stories/**/*.{js,jsx,ts,tsx}' --fix",
     "sb": "storybook dev -p 6006",
     "sb-docs": "storybook dev --docs",
     "sb-build-docs": "storybook build --docs",
@@ -15,12 +15,15 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@stcland/table": "workspace:^",
+    "@stcland/theme": "workspace:^",
     "@storybook/addon-docs": "^8.2.9",
     "@storybook/addon-essentials": "^8.2.9",
     "@storybook/addon-interactions": "^8.2.9",
     "@storybook/addon-links": "^8.2.9",
     "@storybook/react": "^8.2.9",
-    "@vitejs/plugin-react": "^4.3.1",
+    "@tanstack/react-table": "^8.20.5",
+    "@vitejs/plugin-react-swc": "^3.5.0",
     "autoprefixer": "^10.4.20",
     "d3-interpolate": "^3.0.1",
     "postcss": "^8.4.41",
@@ -36,7 +39,6 @@
     "@storybook/theming": "^8.2.9",
     "@tailwindcss/typography": "^0.5.14",
     "@types/d3-interpolate": "^3.0.4",
-    "@types/node": "^20.14.10",
     "@types/react": "^18.3.3",
     "eslint": "^8.56.0",
     "prop-types": "15.8.1",
