+++ packages/storybook/src/index.css
@@ -1,3 +1,1 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
+@import "@stcland/theme/dist/style.css"
