+++ packages/table/src/index.css
@@ -1,1 +1,1 @@
- @import '../../storybook/src/index.css';
+@import "@stcland/theme/dist/style.css"
