+++ packages/theme/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "@stcland/theme",
+  "version": "1.0.0",
+  "description": "",
+  "main": "./dist/index.es.js",
+  "scripts": {
+    "lint": "eslint . --ext .js,.ts 'src/**/*.{js,jsx,ts,tsx}' --fix",
+    "ts-check": "tsc --noEmit --skipLibCheck",
+    "dev": "vite",
+    "build": "vite build"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "autoprefixer": "^10.4.20",
+    "d3-interpolate": "^3.0.1",
+    "postcss": "^8.4.41",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "tailwindcss": "^3.4.10",
+    "typescript": "^5.3.3",
+    "rimraf": "^6.0.1",
+    "eslint": "^8.56.0",
+    "eslint-plugin-react": "^7.35.0"
+  },
+  "devDependencies": {
+    "@tailwindcss/typography": "^0.5.14",
+    "@types/d3-interpolate": "^3.0.4",
+    "@types/react": "^18.3.3",
+    "@vitejs/plugin-react-swc": "^3.5.0",
+    "vite": "^5.4.2",
+    "vite-plugin-css-injected-by-js": "^3.5.1",
+    "vite-plugin-dts": "^4.0.3"
+  }
+}
