+++ packages/theme/src/index.ts
@@ -1,0 +1,6 @@
+
+import './global.css'
+
+export { ThemeProvider, useTheme, themeConfig } from './ThemeProvider'
+
+export type { StclandTheme, TableStyles } from './ThemeProvider'
